pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "utility for constructing a tree of differences between two json files or yaml"
authors = ["pavelkoA <pavelko88@mail.ru>"]
readme = "README.md"
packages = [
    { include = "gendiff"}
]


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"


[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
